+++ core/src/main/java/com/graphhopper/routing/ev/BikeAccessConditional.java
@@ -10,16 +10,16 @@ import java.util.HashSet;
  * Stores temporary so-called conditional restrictions from access:conditional and other conditional
  * tags affecting bikes.
  */
-public enum BikeConditionalAccess {
+public enum BikeAccessConditional {
 
     MISSING, YES, NO;
 
     public static final Collection<String> CONDITIONALS = new HashSet<>(Arrays.asList("access:conditional",
             "vehicle:conditional", "bicycle:conditional"));
-    public static final String KEY = "bike_conditional_access";
+    public static final String KEY = "bike_access_conditional";
 
-    public static EnumEncodedValue<BikeConditionalAccess> create() {
-        return new EnumEncodedValue<>(KEY, BikeConditionalAccess.class);
+    public static EnumEncodedValue<BikeAccessConditional> create() {
+        return new EnumEncodedValue<>(KEY, BikeAccessConditional.class);
     }
 
     @Override