+++ core/src/main/java/com/graphhopper/routing/ev/CarAccessConditional.java
@@ -10,16 +10,16 @@ import java.util.HashSet;
  * Stores temporary so-called conditional restrictions from access:conditional and other conditional
  * tags affecting cars.
  */
-public enum CarConditionalAccess {
+public enum CarAccessConditional {
 
     MISSING, YES, NO;
 
     public static final Collection<String> CONDITIONALS = new HashSet<>(Arrays.asList("access:conditional",
             "vehicle:conditional", "motor_vehicle:conditional", "motorcar:conditional"));
-    public static final String KEY = "car_conditional_access";
+    public static final String KEY = "car_access_conditional";
 
-    public static EnumEncodedValue<CarConditionalAccess> create() {
-        return new EnumEncodedValue<>(KEY, CarConditionalAccess.class);
+    public static EnumEncodedValue<CarAccessConditional> create() {
+        return new EnumEncodedValue<>(KEY, CarAccessConditional.class);
     }
 
     @Override