+++ core/src/main/java/com/graphhopper/routing/ev/FootAccessConditional.java
@@ -10,16 +10,16 @@ import java.util.HashSet;
  * Stores temporary so-called conditional restrictions from access:conditional and other conditional
  * tags affecting foot.
  */
-public enum FootConditionalAccess {
+public enum FootAccessConditional {
 
     MISSING, YES, NO;
 
     public static final Collection<String> CONDITIONALS = new HashSet<>(Arrays.asList("access:conditional",
             "foot:conditional"));
-    public static final String KEY = "foot_conditional_access";
+    public static final String KEY = "foot_access_conditional";
 
-    public static EnumEncodedValue<FootConditionalAccess> create() {
-        return new EnumEncodedValue<>(KEY, FootConditionalAccess.class);
+    public static EnumEncodedValue<FootAccessConditional> create() {
+        return new EnumEncodedValue<>(KEY, FootAccessConditional.class);
     }
 
     @Override