+++ core/src/main/java/com/graphhopper/routing/util/parsers/DefaultTagParserFactory.java
@@ -84,18 +84,18 @@ public class DefaultTagParserFactory implements TagParserFactory {
             return new StateParser(lookup.getEnumEncodedValue(State.KEY, State.class));
         else if (name.equals(Crossing.KEY))
             return new OSMCrossingParser(lookup.getEnumEncodedValue(Crossing.KEY, Crossing.class));
-        else if (name.equals(CarConditionalAccess.KEY)) {
-            EnumEncodedValue<CarConditionalAccess> enc = lookup.getEnumEncodedValue(CarConditionalAccess.KEY, CarConditionalAccess.class);
-            OSMConditionalAccessParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? CarConditionalAccess.YES : CarConditionalAccess.NO);
-            return new OSMConditionalAccessParser(CarConditionalAccess.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
-        } else if (name.equals(BikeConditionalAccess.KEY)) {
-            EnumEncodedValue<BikeConditionalAccess> enc = lookup.getEnumEncodedValue(BikeConditionalAccess.KEY, BikeConditionalAccess.class);
-            OSMConditionalAccessParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? BikeConditionalAccess.YES : BikeConditionalAccess.NO);
-            return new OSMConditionalAccessParser(BikeConditionalAccess.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
-        } else if (name.equals(FootConditionalAccess.KEY)) {
-            EnumEncodedValue<FootConditionalAccess> enc = lookup.getEnumEncodedValue(FootConditionalAccess.KEY, FootConditionalAccess.class);
-            OSMConditionalAccessParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? FootConditionalAccess.YES : FootConditionalAccess.NO);
-            return new OSMConditionalAccessParser(FootConditionalAccess.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
+        else if (name.equals(CarAccessConditional.KEY)) {
+            EnumEncodedValue<CarAccessConditional> enc = lookup.getEnumEncodedValue(CarAccessConditional.KEY, CarAccessConditional.class);
+            OSMAccessConditionalParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? CarAccessConditional.YES : CarAccessConditional.NO);
+            return new OSMAccessConditionalParser(CarAccessConditional.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
+        } else if (name.equals(BikeAccessConditional.KEY)) {
+            EnumEncodedValue<BikeAccessConditional> enc = lookup.getEnumEncodedValue(BikeAccessConditional.KEY, BikeAccessConditional.class);
+            OSMAccessConditionalParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? BikeAccessConditional.YES : BikeAccessConditional.NO);
+            return new OSMAccessConditionalParser(BikeAccessConditional.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
+        } else if (name.equals(FootAccessConditional.KEY)) {
+            EnumEncodedValue<FootAccessConditional> enc = lookup.getEnumEncodedValue(FootAccessConditional.KEY, FootAccessConditional.class);
+            OSMAccessConditionalParser.Setter fct = (edgeId, edgeIntAccess, b) -> enc.setEnum(false, edgeId, edgeIntAccess, b ? FootAccessConditional.YES : FootAccessConditional.NO);
+            return new OSMAccessConditionalParser(FootAccessConditional.CONDITIONALS, fct, properties.getString("date_range_parser_day", ""));
         } else if (name.equals(FerrySpeed.KEY))
             return new FerrySpeedCalculator(lookup.getDecimalEncodedValue(FerrySpeed.KEY));
         return null;